death alteration approal list by pin updated

+++ src/main/java/lk/rgd/crs/core/service/DeathAlterationServiceImpl.java
@@ -3,15 +3,15 @@ package lk.rgd.crs.core.service;
 import lk.rgd.crs.api.service.DeathAlterationService;
 import lk.rgd.crs.api.service.DeathRegistrationService;
 import lk.rgd.crs.api.domain.DeathAlteration;
-import lk.rgd.crs.api.domain.BDDivision;
 import lk.rgd.crs.api.domain.DeathRegister;
 import lk.rgd.crs.api.dao.DeathAlterationDAO;
 import lk.rgd.crs.CRSRuntimeException;
+import lk.rgd.crs.core.ValidationUtils;
 import lk.rgd.crs.web.WebConstants;
 import lk.rgd.common.api.domain.User;
 import lk.rgd.common.api.domain.Role;
-import lk.rgd.common.api.domain.DSDivision;
 import lk.rgd.ErrorCodes;
+import lk.rgd.Permission;
 import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,7 +24,7 @@ import java.util.*;
  */
 public class DeathAlterationServiceImpl implements DeathAlterationService {
 
-    private static final Logger logger = LoggerFactory.getLogger(BirthAlterationServiceImpl.class);
+    private static final Logger logger = LoggerFactory.getLogger(DeathAlterationServiceImpl.class);
     private final DeathAlterationDAO deathAlterationDAO;
     private final DeathRegistrationService deathRegistrationService;
     private final DeathAlterationValidator deathAlterationValidator;
@@ -62,6 +62,7 @@ public class DeathAlterationServiceImpl implements DeathAlterationService {
     @Transactional(propagation = Propagation.REQUIRED)
     public void deleteDeathAlteration(long idUKey, User user) {
         logger.debug("about to remove alteration recode idUkey : {}", idUKey);
+        validateAccessOfUserToEditOrDelete(getById(idUKey, user), user);
         deathAlterationDAO.deleteDeathAlteration(idUKey);
     }
 
@@ -72,6 +73,7 @@ public class DeathAlterationServiceImpl implements DeathAlterationService {
         DeathAlteration exsisting = getById(idUKey, user);
         exsisting.setStatus(DeathAlteration.State.REJECT);
         exsisting.setComments(comment);
+        validateAccessOfUserToEditOrDelete(exsisting, user);
         deathAlterationDAO.rejectDeathAlteration(exsisting, user);
     }
 
@@ -96,7 +98,7 @@ public class DeathAlterationServiceImpl implements DeathAlterationService {
      * @inheritDoc
      */
     @Transactional(propagation = Propagation.REQUIRED)
-    public List<DeathAlteration> getAlterationApprovalListByDeathDivision(int pageNo, int numRows, int divisionId) {
+    public List<DeathAlteration> getAlterationApprovalListByDeathDivision(int pageNo, int numRows, int divisionId, User user) {
         return deathAlterationDAO.getPaginatedAlterationApprovalListByDeathDivision(pageNo, numRows, divisionId);
     }
 
@@ -155,7 +157,7 @@ public class DeathAlterationServiceImpl implements DeathAlterationService {
      * @inheritDoc
      */
     @Transactional(propagation = Propagation.REQUIRED)
-    public List<DeathAlteration> getDeathAlterationByUserLocation(int locationUKey) {
+    public List<DeathAlteration> getDeathAlterationByUserLocation(int locationUKey, User user) {
         List<DeathAlteration> result = deathAlterationDAO.getDeathAlterationByUserLocation(locationUKey);
         Iterator itr = result.iterator();
         while (itr.hasNext()) {
@@ -176,30 +178,57 @@ public class DeathAlterationServiceImpl implements DeathAlterationService {
         while (itr.hasNext()) {
             DeathAlteration da = (DeathAlteration) itr.next();
             loadValuesToDeathAlterationObject(da);
+            validateAccessOfUserForApproval(da, user);
         }
         return result;
     }
 
-
-/*    private void validateAccessToBDDivision(User user, BDDivision bdDivision) {
-        if (!(User.State.ACTIVE == user.getStatus()
-                &&
-                (Role.ROLE_ARG.equals(user.getRole().getRoleId())
-                        ||
-                        (user.isAllowedAccessToBDDistrict(bdDivision.getDistrict().getDistrictUKey())
-                                &&
-                                user.isAllowedAccessToBDDSDivision(bdDivision.getDsDivision().getDsDivisionUKey())
-                        )
-                )
-        )) {
-                       handleException("User : " + user.getUserId() + " is not allowed access to the District : " +
-                    bdDivision.getDistrict().getDistrictId() + " and/or DS Division : " +
-                    bdDivision.getDsDivision().getDivisionId(), ErrorCodes.PERMISSION_DENIED);
+    /**
+     * Checks if the user can edit or delete a death alteration entry before approval by an ARG
+     * <p/>
+     * A DEO or ADR at the same "SubissionLocation" can edit or delete an entry at data entry stage. Any other who has
+     * access to the BD division of the corresponding BDF has access to edit or delete
+     *
+     * @param da   the death alteration entry
+     * @param user the user attempting to update or delete
+     */
+    private void validateAccessOfUserToEditOrDelete(DeathAlteration da, User user) {
+        if (Role.ROLE_DEO.equals(user.getRole().getRoleId()) || Role.ROLE_ADR.equals(user.getRole().getRoleId())) {
+            if (da.getSubmittedLocation().equals(user.getPrimaryLocation())) {
+                return;
+            }
+        } else if (!Role.ROLE_ADMIN.equals(user.getRole().getRoleId())) {
+            return;
         }
-    }*/
+
+        ValidationUtils.validateAccessToBDDivision(user, deathAlterationDAO.getById(da.getDeathId()).getDeathRecodDivision());
+    }
+
+    /**
+     * Checks if the user can approve the death alteration entry. Only an ARG level user, in charge of the BDF under
+     * consideration can approve an alteration
+     *
+     * @param da   the death alteration entry
+     * @param user the user attempting to approve
+     */
+    private void validateAccessOfUserForApproval(DeathAlteration da, User user) {
+        if (Role.ROLE_RG.equals(user.getRole().getRoleId())) {
+            // RG can approve any record
+        } else if (Role.ROLE_ARG.equals(user.getRole().getRoleId())) {
+            ValidationUtils.validateAccessToBDDivision(user, deathAlterationDAO.getById(da.getDeathId()).getDeathRecodDivision());
+            if (!user.isAuthorized(Permission.APPROVE_BIRTH_ALTERATION)) {
+                handleException("User : " + user.getUserId() + " is not allowed to approve/reject death alteration",
+                        ErrorCodes.PERMISSION_DENIED);
+            }
+        } else {
+            handleException("User : " + user.getUserId() + " is not an ARG for alteration approval",
+                    ErrorCodes.PERMISSION_DENIED);
+        }
+    }
 
     private void handleException(String message, int code) {
         logger.error(message);
         throw new CRSRuntimeException(message, code);
     }
+
 }
